backend/requirements.txt
# Web framework and server
fastapi>=0.68.0
uvicorn>=0.15.0

# Database
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0

# API utilities
pydantic>=1.8.2
python-jose>=3.3.0
python-multipart>=0.0.5

# ML/NLP dependencies - using specific versions known to work well together
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=1.9.0
transformers>=4.11.0
spacy>=3.1.0
